requirements.txt
Flask
requests
pandas
gunicorn
SQLAlchemy
Werkzeug
python-dotenv
psycopg2-binary